\name{AteMeVs-package}
\alias{AteMeVs-package}
\docType{package}
\title{
Estimation of average treatment effects with measurement error and variable selection for confounders
}
\description{
A recent method proposed by Yi and Chen (2023) <doi:10.1177/09622802221146308> is implemented to estimate the average treatment effects using noisy data containing both measurement error and spurious variables.
}
\details{
The R package 'AteMeVs', which refers to estimation of the Average Treatment Effects with Measurement Error and Variable Selection for confounders, contains a set of functions that provide a step-by-step estimation procedure, including the correction of the measurement error effects, variable selection for building the model used to estimate the propensity scores, and estimation of the average treatment effects. The functions contain multiple options for users to implement, including different ways to correct for the measurement error effects, distinct choices of penalty functions to do variable selection, and various regression models to characterize propensity scores.
}

\author{
Chen, L.-P. and Yi, G. Y.

Maintainer: Li-Pang Chen <lchen723@nccu.edu.tw>
}
\references{
Yi, G. Y. and Chen, L.-P. (2023). Estimation of the average treatment effect with variable selection and measurement error simultaneously addressed for potential confounders. \emph{Statistical Methods in Medical Research}, 32, 691-711.
}


\keyword{ package }

